+++ lib/maze/hooks/appium_hooks.rb
@@ -38,10 +38,7 @@ module Maze
         if Maze.config.os == 'macos'
           # Close the app - without the sleep, launching the app for the next scenario intermittently fails
           system("killall -KILL #{Maze.config.app} && sleep 1")
-        elsif [:bb].include? Maze.config.farm
-          Maze.driver.terminate_app Maze.driver.app_id
-          Maze.driver.activate_app Maze.driver.app_id
-        else
+        elsif [:bb, :bs, :local].include? Maze.config.farm
           Maze.driver.terminate_app Maze.driver.app_id
           Maze.driver.activate_app Maze.driver.app_id
         end