+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Add step to sort received requests by specific field [402](https://github.com/bugsnag/maze-runner/pull/402)
 
+## Fixes
+
+- Fix local mode, broken in v7 refactor [403](https://github.com/bugsnag/maze-runner/pull/403)
+
 # 7.1.0 - 2022/09/30
 
 ## Enhancements