+++ CHANGELOG.md
@@ -4,6 +4,16 @@
 
 - Add `/trace` endpoint [402](https://github.com/bugsnag/maze-runner/pull/402)
 
+# 7.2.0 - 2022/10/04
+
+## Enhancements
+
+- Add step to sort received requests by specific field [402](https://github.com/bugsnag/maze-runner/pull/402)
+
+## Fixes
+
+- Fix local mode, broken in v7 refactor [403](https://github.com/bugsnag/maze-runner/pull/403)
+
 # 7.1.0 - 2022/09/30
 
 ## Enhancements